add module for spec_helper
